move pages and other folders in src
