+++ web/src/mocks/averagePrices.ts
