+++ web/src/modules/home/components/SquareGuides/Rectangle/OneMiniSquare/index.tsx
