+++ web/src/modules/home/components/SquareGuides/Rectangle/index.tsx
