+++ web/src/modules/search/components/AveragePrices/TableBody/index.tsx
