+++ web/src/modules/search/components/AveragePrices/TableHead/index.tsx
