+++ web/src/modules/search/components/AveragePrices/index.tsx
