+++ web/src/modules/search/components/BestOption/OneVariant/index.tsx
