+++ web/src/modules/search/components/FilterPage/FeaturesAmenities/index.tsx
