+++ web/src/modules/search/components/FilterPage/MonthlyPriceRange/index.tsx
