+++ web/src/modules/search/components/Map/index.tsx
