+++ web/src/modules/search/components/ViewOnMap/index.tsx
