+++ web/src/pages/_app.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import Head from 'next/head';
 import { ThemeProvider } from '@material-ui/core/styles';
 import CssBaseline from '@material-ui/core/CssBaseline';
-import { MuiTheme } from "../src/MuiTheme"
+import { MuiTheme } from "../MuiTheme"
 
 type Props = {
     Component:any,
